Workaround doors 'bug' with entity twins and 'events' cascade
Previously, doors could be only opened on one side (the one actually
containing that Entity::Door). I could have taken into account entities
on the other side of a wall, but that would probably cause future
issues: you don't necesarily want *any* entity on edge to be displayed
on both sides, such as paintings ; doors are a special case and I think
this approach is the best, ie. handling this on the lowest level
possible (on the data layer rather than the UI).